digital_twin_for_crop_growth_server/requirements.txt
Django==4.1.7
djangorestframework==3.15.1
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
PyMySQL==1.1.0
torch==2.2.2
